Wrapping "MyAccount" - Login
Added validation for fields: Username and Email on create account page.

Added action in form for login.

Added /login and "/fonts/**"-folder as public matchers  in
SecurityConfig, that user can access now.

Added validation for fields: Username and password in Login form.

Added validation for forgetPassword and modify metohd forgetPassword().

Added validation for forgetPassword. But we have a litle issue with
forgetPassword.